refactor: move pool helper to src/internal
