+++ src/clients/main.client.ts
@@ -1,6 +1,6 @@
 import { type CacheStore, EtagStore, MemoryCache } from "../config/cache";
+import { DEFAULT_CONCURRENCY, mapWithConcurrency } from "../internal/pool";
 import type { APIResource, NamedAPIResource } from "../models/common/resource";
-import { DEFAULT_CONCURRENCY, mapWithConcurrency } from "../utils/pool";
 import type { ClientOptions, RequestScope } from "./base";
 import { BerryClient } from "./berry.client";
 import { ContestClient } from "./contest.client";
