+++ src/internal/pool.ts
